Refactor: Stack MVP 검증 수행 단계 수정 및 프론트에서 alert 창 띄우기

+++ back/src/routers/stackRouter.js
@@ -10,6 +10,10 @@ stackRouter.get('/', async(req,res,next)=>{
 
     try{
         const stacks = await stackService.findAll({ user_id });
+        if (!stacks){
+            res.status(400).send({error: '유저의 스택 정보가 존재하지 않습니다.'})
+            throw new Error(`${user_id} 유저의 스택 정보가 존재하지 않습니다.`);
+        }
         res.status(200).json(stacks)
     }catch(error){
         next(error);
@@ -20,15 +24,26 @@ stackRouter.get('/', async(req,res,next)=>{
 stackRouter.post('/', async(req, res, next) => {
     try{
         if(is.emptyObject(req.body)){
+            res.status(400).send(error);
             throw new Error('headers의 Content-Type을 application/json으로 설정해주세요');
         }
 
         const user_id = req.currentUserId;
-        const {stackName, stackDescription} = req.body;
 
+        const stacks = await stackService.findAll({ user_id });
+
+        const {stackName, stackDescription} = req.body;
         const newStack = {stackName, stackDescription};
+        
+        const stackExists = stacks.some((stack) => stack.stackName === newStack.stackName);
+        if (stackExists ){
+            res.status(400).send({error: `${newStack.stackName} 프로젝트는 이미 존재합니다.`})
+            throw new Error(`${newStack.stackName} 프로젝트는 이미 존재합니다.`);
+        }
+        
 
         if(!stackName || !stackDescription){
+            res.status(400).send({error: '모든 값을 입력했는지 확인해주세요.'})
             throw new Error('모든 값을 입력했는지 확인해주세요.');
         }
 
@@ -53,10 +68,23 @@ stackRouter.put('/:stack_id', async(req, res, next) => {
         const { stackName, stackDescription } = req.body;
 
         if(!stackName || !stackDescription){
+            res.status(400).send({error: '모든 값을 입력했는지 확인해주세요.'});
             throw new Error('모든 값을 입력했는지 확인해주세요.');
         }
 
         const newStack = {stackName, stackDescription};
+        const stack = await stackService.findOne({stack_id});
+        if (!stack) {
+            res.status(400).send({error: '이 스택 정보는 존재하지 않습니다.'});
+            throw new Error('이 스택 정보는 존재하지 않습니다.');
+        }
+        // 수정하려는 스택만 제외하고 다른 모든 스택을 가져올 수 있는 방법
+        const exceptStacks = await stackService.findExcept({ user_id, stack_id });
+        const stackExists = exceptStacks.some((stack) => stack.stackName === newStack.stackName);
+        if (stackExists ){
+            res.status(400).send({error: `${newStack.stackName} 프로젝트는 이미 존재합니다.`})
+            throw new Error(`${newStack.stackName} 프로젝트는 이미 존재합니다.`);
+        }
         const updatedStack = await stackService.updateStack({ user_id, stack_id, newStack });
         res.status(200).json(updatedStack);
     }catch(error){
@@ -70,6 +98,12 @@ stackRouter.delete('/:stack_id', async(req,res,next) => {
         const user_id = req.currentUserId;
         const { stack_id } = req.params;
 
+        const stack = await stackService.findOne({stack_id});
+        if(!stack){
+            res.status(400).send({error: '이 스택 정보는 존재하지 않습니다.'})
+            throw new Error('이 스택 정보는 존재하지 않습니다.');
+        }
+        
         const deletedStack = await stackService.deleteStack({ user_id, stack_id });
         res.status(200).json(deletedStack);
